moving code from current Main.java to a new Exercise1.java file
